+++ src/rider/main/kotlin/com/github/rafaelldi/diagnosticsclientplugin/model/DiagnosticsHostModel.kt
@@ -146,6 +146,8 @@ object DiagnosticsHostModel : Ext(DiagnosticsHostRoot) {
         +"GcHeapSize"
         +"Gc"
         +"Exception"
+        +"ExceptionCount"
+        +"ThreadCount"
     }
 
     init {